+++ src/main/java/com/bibireden/data_attributes/api/event/EntityAttributeModifiedEvents.java
@@ -1,5 +1,6 @@
 package com.bibireden.data_attributes.api.event;
 
+import net.minecraft.registry.entry.RegistryEntry;
 import org.jetbrains.annotations.Nullable;
 
 import net.fabricmc.fabric.api.event.Event;
@@ -21,9 +22,9 @@ public final class EntityAttributeModifiedEvents {
 	 * the value of the modifier, or by reloading the datapack and having the living entity renew its attribute container. 
 	 * Living entity and modifiers may or may not be null.
 	 */
-	public static final Event<Modified> MODIFIED = EventFactory.createArrayBacked(Modified.class, callbacks -> (attribute, livingEntity, modifier, prevValue, isWasAdded) -> {
+	public static final Event<Modified> MODIFIED = EventFactory.createArrayBacked(Modified.class, callbacks -> (entry, livingEntity, modifier, prevValue, isWasAdded) -> {
 		for(Modified callback : callbacks) {
-			callback.onModified(attribute, livingEntity, modifier, prevValue, isWasAdded);
+			callback.onModified(entry, livingEntity, modifier, prevValue, isWasAdded);
 		}
 	});
 	
@@ -31,11 +32,11 @@ public final class EntityAttributeModifiedEvents {
 	 * Fired after the attribute instance value was calculated, but before it was output. This offers one last chance to alter the 
 	 * value in some way (for example round a decimal to an integer).
 	 */
-	public static final Event<Clamped> CLAMPED = EventFactory.createArrayBacked(Clamped.class, callbacks -> (attribute, value) -> {
+	public static final Event<Clamped> CLAMPED = EventFactory.createArrayBacked(Clamped.class, callbacks -> (entry, value) -> {
 		double cache = value;
 		
 		for(Clamped callback : callbacks) {
-			cache = callback.onClamped(attribute, cache);
+			cache = callback.onClamped(entry, cache);
 		}
 		
 		return cache;
@@ -43,11 +44,11 @@ public final class EntityAttributeModifiedEvents {
 	
 	@FunctionalInterface
 	public interface Modified {
-		void onModified(final EntityAttribute attribute, final @Nullable LivingEntity livingEntity, final @Nullable EntityAttributeModifier modifier, final double prevValue, final boolean isWasAdded);
+		void onModified(final RegistryEntry<EntityAttribute> entry, final @Nullable LivingEntity livingEntity, final @Nullable EntityAttributeModifier modifier, final double prevValue, final boolean isWasAdded);
 	}
 	
 	@FunctionalInterface
 	public interface Clamped {
-		double onClamped(final EntityAttribute attribute, final double value);
+		double onClamped(final RegistryEntry<EntityAttribute> entry, final double value);
 	}
 }