+++ src/main/java/com/bibireden/data_attributes/api/item/ItemFields.java
@@ -3,9 +3,10 @@ package com.bibireden.data_attributes.api.item;
 import java.util.UUID;
 
 import net.minecraft.item.Item;
+import net.minecraft.util.Identifier;
 
 /**
- * Helper class that exposes {@link Item#ATTACK_DAMAGE_MODIFIER_ID} and {@link Item#ATTACK_SPEED_MODIFIER_ID}.
+ * Helper class that exposes {@link Item#BASE_ATTACK_DAMAGE_MODIFIER_ID} and {@link Item#BASE_ATTACK_SPEED_MODIFIER_ID}.
  * 
  * @author CleverNucleus
  *
@@ -13,7 +14,7 @@ import net.minecraft.item.Item;
 public final class ItemFields extends Item {
 	private ItemFields() { super(new Item.Settings()); }
 	
-	public static UUID attackDamageModifierID() { return ATTACK_DAMAGE_MODIFIER_ID; }
+	public static Identifier attackDamageModifierID() { return BASE_ATTACK_DAMAGE_MODIFIER_ID; }
 	
-	public static UUID attackSpeedModifierID() { return ATTACK_SPEED_MODIFIER_ID; }
+	public static Identifier attackSpeedModifierID() { return BASE_ATTACK_SPEED_MODIFIER_ID; }
 }