+++ src/main/kotlin/com/bibireden/data_attributes/DataAttributesClient.kt
@@ -9,8 +9,7 @@ import net.fabricmc.api.ClientModInitializer
 import net.fabricmc.fabric.api.client.networking.v1.ClientLoginNetworking
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking
 import net.fabricmc.fabric.api.networking.v1.PacketByteBufs
-import net.minecraft.client.MinecraftClient
-import net.minecraft.network.PacketByteBuf
+import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry
 import java.util.concurrent.CompletableFuture
 
 class DataAttributesClient : ClientModInitializer {
@@ -19,13 +18,9 @@ class DataAttributesClient : ClientModInitializer {
         val MANAGER = AttributeConfigManager()
 
         /** Whenever the client receives a sync packet from the server to update the world-state via. configuration. */
-        fun onPacketReceived(client: MinecraftClient, buf: PacketByteBuf) {
-            buf.retain()
-            client.execute {
-                MANAGER.readPacket(AttributeConfigManager.Packet.ENDEC.decodeFully(ByteBufDeserializer::of, buf))
-                MANAGER.onDataUpdate()
-                buf.release()
-            }
+        fun onPacketReceived(packet: AttributeConfigManager.Packet) {
+            MANAGER.readPacket(packet)
+            MANAGER.onDataUpdate()
         }
     }
 
@@ -33,11 +28,12 @@ class DataAttributesClient : ClientModInitializer {
         ConfigScreen.registerProvider(DataAttributes.MOD_ID) { DataAttributesConfigScreen(DataAttributes.OVERRIDES_CONFIG, DataAttributes.FUNCTIONS_CONFIG, DataAttributes.ENTITY_TYPES_CONFIG, it) }
 
         ClientLoginNetworking.registerGlobalReceiver(NetworkingChannels.HANDSHAKE) { client, _, buf, _ ->
-            onPacketReceived(client, buf)
+            AttributeConfigManager.Packet.ENDEC.decodeFully(ByteBufDeserializer::of, buf)?.let(::onPacketReceived)
             CompletableFuture.completedFuture(PacketByteBufs.empty())
         }
-        ClientPlayNetworking.registerGlobalReceiver(NetworkingChannels.RELOAD) { client, _, buf, _ ->
-            onPacketReceived(client, buf)
+
+        ClientPlayNetworking.registerGlobalReceiver(AttributeConfigManager.Packet.PACKET_ID) { packet, _ ->
+            onPacketReceived(packet)
         }
     }
 }