+++ src/main/kotlin/com/bibireden/data_attributes/networking/ConfigPacketBufs.kt
@@ -6,17 +6,14 @@ import com.bibireden.data_attributes.data.EntityTypeData
 import io.wispforest.endec.Endec
 import io.wispforest.endec.format.bytebuf.ByteBufDeserializer
 import io.wispforest.endec.format.bytebuf.ByteBufSerializer
-import io.wispforest.owo.network.serialization.PacketBufSerializer
+import io.wispforest.endec.impl.ReflectiveEndecBuilder
 import org.jetbrains.annotations.ApiStatus
 import kotlin.reflect.KClass
 
 /** All serializers/deserializers related to [ByteBufSerializer] for the Data Attributes config. */
 @ApiStatus.Internal
 object ConfigPacketBufs {
-    private fun <T : Any> registerSerializer(clazz: KClass<T>, endec: Endec<T>) = PacketBufSerializer.register(clazz.java,
-        { buf, t -> endec.encodeFully({ ByteBufSerializer.of(buf) }, t); return@register },
-        { buf -> endec.decodeFully(ByteBufDeserializer::of, buf) }
-    )
+    private fun <T : Any> registerSerializer(clazz: KClass<T>, endec: Endec<T>) = ReflectiveEndecBuilder.SHARED_INSTANCE.register(endec, clazz.java)
 
     fun registerPacketSerializers() {
         registerSerializer(AttributeOverride::class, AttributeOverride.ENDEC)