+++ src/main/kotlin/com/bibireden/data_attributes/networking/NetworkingChannels.kt
@@ -1,12 +1,13 @@
 package com.bibireden.data_attributes.networking
 
 import com.bibireden.data_attributes.DataAttributes
+import net.minecraft.network.packet.CustomPayload.Id
 import org.jetbrains.annotations.ApiStatus
 
 @ApiStatus.Internal
 object NetworkingChannels {
     @JvmField
-    val HANDSHAKE = DataAttributes.id("handshake")
+    val HANDSHAKE = (DataAttributes.id("handshake"))
     @JvmField
     val RELOAD = DataAttributes.id("reload")
 }