+++ CHANGELOG.md
@@ -1,2 +1,7 @@
-## Changes ⚙️
-- Now using `WrapMethod` on computation of modifier(s) to avoid unneeded recalculation.
+## Changes 🌽
+- Updated dependencies to `1.21` equivalent versions.
+- Changed internal networking to abide by version.
+- `ItemFields` have changed to fit `1.21` definitions.
+- 💖 The API now takes in `RegistryEntry` along with the targeted attribute.
+    - This also goes for **all events**.
+    - `IEntityAttribute` has also been extended to get its `RegistryEntry`.